Add files via upload
This is the database diagram created in Vertabelo which shows all the tables, their relationships and the Airplane design as a whole. The SQL code was generated from this diagram and then edited to follow a particular coding standard and to add own functions, stored procedures, views and also data entry.